Stats done - implemented with binomial trial & Count trial:

+++ app/src/main/java/com/example/nerdherd/MeasurementTrial.java
@@ -12,7 +12,7 @@ public class MeasurementTrial extends Trial {
         return totalMeasurementCount;
     }
 
-    public int getTotalMeasurementCount() {
+    public double getTotalMeasurementCount() {
         return totalMeasurementCount;
     }
